requirements.txt
# requirements.txt
requests>=2.26.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.5.0